+++ CHANGELOG.md
@@ -41,6 +41,8 @@ IMPROVEMENTS:
   - providers/virtualbox: customizations via VBoxManage are retried, avoiding
     VirtualBox flakiness [GH-2483]
   - provisioners/ansible: allow files for extra vars [GH-2366]
+  - provisioners/puppet: client cert and private key can now be specified
+    for the puppet server provisioner. [GH-902]
   - provisioners/shell: Added `keep_color` option to not automatically color
     output based on stdout/stderr. [GH-2505]
   - provisioners/shell: Arguments can now be an array of args. [GH-1949]